+++ types/router.go
@@ -2,6 +2,6 @@ package types
 
 type RouterConfig struct {
 	Providers []ProviderConfig
-	MaxTokens int
+	MaxTokens int64
 	Model     string
 }
